bcoinapp: separate txinput from bcoin app
