+++ lib/bledger.js
@@ -7,11 +7,14 @@
 'use strict';
 
 const hid = require('./ledgerhid');
+
 const APDU = require('./apdu');
-const {SignInput, LedgerBcoin} = require('./bcoin');
-const BTC = require('./btc');
 const LedgerError = require('./error');
 const LedgerProtocol = require('./ledgerprotocol');
+
+const BTC = require('./btc');
+const LedgerBcoin = require('./bcoin');
+const LedgerTXInput = require('./txinput');
 
 exports.bledger = exports;
 
@@ -20,7 +23,7 @@ exports.LedgerProtocol = LedgerProtocol;
 exports.LedgerError = LedgerError;
 
 exports.LedgerBcoin = LedgerBcoin;
-exports.SignInput = SignInput;
+exports.LedgerTXInput = LedgerTXInput;
 
 exports.LedgerBTC = BTC;
