+++ test/data/tx-p2sh-mulsig.json
@@ -1,6 +1,6 @@
 {
   "tx": "01000000025206bb6a032940b6a5944d62c219db675e44db8328f4b13cdb7f53e0ecc518630000000000ffffffffc2d4431b956602b7a8e86faf15c53861d6971e33929e4e2a65a751c31baa01890000000000ffffffff0200e1f5050000000017a9146de7cfcbfe6218ca91470958d0e9294880444d4d87b1bdf5050000000017a9145ec134e75e148fcf0413b09de480f2f9ac5868528700000000",
-  "signInputs": [
+  "ledgerInputs": [
     {
       "tx": "0100000001f6f2cd95c6eafee1707a5a46f2d54c14501fe92aab7e2358339e0b2d75e3b2be000000006b4830450221009fa1718346e9269353a93e04169947b138f367bb8fa67bc8e31ecf21dba287600220792582d62e4d91ec89daae3853e6c0d611faf9b550a5f53b39f1fb246500aed4012102203f1ed4c881ed19939f7d6c4ec329c22b69b55bd6a547abfc9ceec363493f8cffffffff0137d8f5050000000017a9145ec134e75e148fcf0413b09de480f2f9ac5868528700000000",
       "index": 0,
