+++ test/data/tx2.json
@@ -1,7 +1,7 @@
 {
   "tx": "0100000003ae29b8f0086276bb51a4abe8c107cfa44500bbd824f15a1beb86a0d72b27b1d00100000000ffffffff1a0315868daab16de904de31821e5e3ed683a4d5b4b299ae62ab416571c85ff50000000000ffffffff6481cf32f63febafbce71e3802a2c139bece8964eb3789202a5d7d892c582d4d0000000000ffffffff02601ce011000000001976a9145e579a42b98e046e65de2b41b227f7072a36a32488ac00000000000000000f6a0d48656c6c6f20626c656467657200000000",
   "signedTX": "0100000003ae29b8f0086276bb51a4abe8c107cfa44500bbd824f15a1beb86a0d72b27b1d0010000006a4730440220459e6c7d09160a36ecaa65f4d5fd6c7227b1e27a4d6bff28563318009aacdd0a02201f4ee875e60a1aaa18169a8f3cca4e7de89f35733da65097b4b8f1cf9686e7b0012103180be38b6b2aa5edb19c1f6c374bfc7f07bf5c929a9513c5526d668a2a06f2acffffffff1a0315868daab16de904de31821e5e3ed683a4d5b4b299ae62ab416571c85ff5000000006b483045022100ba6ba1d25339f634cd2ecfd5602c7413e5ac10d5b9b68e7d9386810e3ebbc19f022062ac7788705328753a0252c011f7f896a5bdd9ed6ac022d7349e5fa396bd0ed60121032cd1febba4076743a09009e6ac9abb2d34a66f188b1bc303d06164ad6ccc62cbffffffff6481cf32f63febafbce71e3802a2c139bece8964eb3789202a5d7d892c582d4d000000006a4730440220213d424acd2dde0a5e5d611edc54838675535f63a9befa013fc7167fd911da7102204d463518827d0615868373e17c63aa68720e11a7805e66d9792428f6dcb22f320121032cd1febba4076743a09009e6ac9abb2d34a66f188b1bc303d06164ad6ccc62cbffffffff02601ce011000000001976a9145e579a42b98e046e65de2b41b227f7072a36a32488ac00000000000000000f6a0d48656c6c6f20626c656467657200000000",
-  "signInputs": [
+  "ledgerInputs": [
     {
       "tx": "01000000011674edca60ce7863a3bbf04fdaad83aa9bbbd90d57f269134aca8bedae5355b6010000006b483045022100d03936b5856d76952243535793476b60d83223a111167bec2db98368008dd21e0220313d824bc7cd470066423aabc91aad5bacd39205eb4cf5038916311d1e0159fb012103e98a09a729418ed0cd56c01559e3f317a522b3f24027cc455ea2bc09759567e4ffffffff0280e641ac000000001976a914431321a30e44eaaf87bcc0fd134852bf73b0a52388ac00e1f505000000001976a914efd047c8d895b76cbdcd84a798695e76a0fdb3e488ac00000000",
       "index": 1,
